database/requirements.txt
# 自动化数据管道依赖

# 核心依赖
pandas>=1.5.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
requests>=2.28.0

# 任务调度
APScheduler>=3.10.0

# 通知服务
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0

# 邮件支持 (可选)
# 如果需要发送邮件通知，取消注释
# secure-smtplib>=0.1.1

# 开发工具 (可选)
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
